🏗️ Architect: Modularized planets search
Refactored the 509-line `apts/skyfield_searches/planets.py` 'God Object' into a modular package `apts/skyfield_searches/planets/`.

Changes:
- Created `apts/skyfield_searches/planets/` package.
- Decomposed logic into:
    - `extrema.py`: Observational maxima/minima (Dichotomy, Greatest Elongations, Stationary Points, Greatest Brilliancy, Highest Altitude).
    - `orbits.py`: Orbital events (Aphelion/Perihelion, Oppositions, Mars Closest Approach).
    - `alignments.py`: Multi-planet alignments.
- Maintained public API via `__init__.py` re-exports.
- Updated relative imports to account for the new package structure.

This improvement follows the Single Responsibility Principle and enhances maintainability by isolating distinct planetary search algorithms.
